refactor(cli): extract command registration
